requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.5.0
google-cloud-logging==3.9.0
google-cloud-firestore==2.16.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
pytest==8.1.0
